crawling/requirements.txt
PyMySQL==1.1.1
python-dotenv==1.0.1
Requests==2.32.3
schedule==1.2.2
pytz==2022.7.1
cryptography==39.0.1
python-dateutil
selenium
